A little refactoring Moved grid related things to package net.rptools.maptool.model.grid Introduced some enums and classes to support interesting grids
