+++ src/main/java/net/rptools/maptool/client/functions/ZoomFunctions.java
@@ -21,8 +21,8 @@ import net.rptools.maptool.client.AppState;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.model.CellPoint;
-import net.rptools.maptool.model.Grid;
 import net.rptools.maptool.model.ZonePoint;
+import net.rptools.maptool.model.grid.Grid;
 import net.rptools.maptool.util.FunctionUtil;
 import net.rptools.parser.Parser;
 import net.rptools.parser.ParserException;
