+++ src/main/java/net/rptools/maptool/client/functions/getInfoFunction.java
@@ -36,8 +36,6 @@ import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.Campaign;
 import net.rptools.maptool.model.CampaignProperties;
-import net.rptools.maptool.model.Grid;
-import net.rptools.maptool.model.GridFactory;
 import net.rptools.maptool.model.Light;
 import net.rptools.maptool.model.LightSource;
 import net.rptools.maptool.model.LookupTable;
@@ -46,6 +44,8 @@ import net.rptools.maptool.model.Token;
 import net.rptools.maptool.model.Zone;
 import net.rptools.maptool.model.drawing.DrawableColorPaint;
 import net.rptools.maptool.model.drawing.DrawableTexturePaint;
+import net.rptools.maptool.model.grid.Grid;
+import net.rptools.maptool.model.grid.GridFactory;
 import net.rptools.maptool.server.ServerPolicy;
 import net.rptools.maptool.util.FunctionUtil;
 import net.rptools.maptool.util.MapToolSysInfoProvider;
