+++ src/main/java/net/rptools/maptool/client/ui/mappropertiesdialog/MapPropertiesDialog.java
@@ -38,15 +38,15 @@ import net.rptools.maptool.client.ui.theme.RessourceManager;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.Asset;
 import net.rptools.maptool.model.AssetManager;
-import net.rptools.maptool.model.Grid;
-import net.rptools.maptool.model.GridFactory;
-import net.rptools.maptool.model.GridlessGrid;
-import net.rptools.maptool.model.HexGridHorizontal;
-import net.rptools.maptool.model.HexGridVertical;
-import net.rptools.maptool.model.IsometricGrid;
-import net.rptools.maptool.model.SquareGrid;
 import net.rptools.maptool.model.Zone;
 import net.rptools.maptool.model.drawing.DrawablePaint;
+import net.rptools.maptool.model.grid.Grid;
+import net.rptools.maptool.model.grid.GridFactory;
+import net.rptools.maptool.model.grid.GridlessGrid;
+import net.rptools.maptool.model.grid.HexGridHorizontal;
+import net.rptools.maptool.model.grid.HexGridVertical;
+import net.rptools.maptool.model.grid.IsometricGrid;
+import net.rptools.maptool.model.grid.SquareGrid;
 import net.rptools.maptool.util.ImageManager;
 import net.rptools.maptool.util.StringUtil;
