--- src/main/java/net/rptools/maptool/model/GridDrawBlendComposite.java
@@ -1,38 +1,0 @@
-/*
- * This software Copyright by the RPTools.net development team, and
- * licensed under the Affero GPL Version 3 or, at your option, any later
- * version.
- *
- * MapTool Source Code is distributed in the hope that it will be
- * useful, but WITHOUT ANY WARRANTY; without even the implied warranty
- * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- *
- * You should have received a copy of the GNU Affero General Public
- * License * along with this source Code.  If not, please visit
- * <http://www.gnu.org/licenses/> and specifically the Affero license
- * text at <http://www.gnu.org/licenses/agpl.html>.
- */
-package net.rptools.maptool.model;
-
-public enum GridDrawBlendComposite {
-  NONE,
-  /*  potential future blend modes
-   ADD,
-   BLUE,
-   BURN,
-   DODGE,
-   DARKEN,
-   DIFFERENCE,
-   EXCLUSION,
-   GREEN,
-   HARD_LIGHT,
-   LIGHTEN,
-   MULTIPLY,
-   OVERLAY,
-   RED,
-   SCREEN,
-   SOFT_LIGHT,
-   SRC_ATOP,
-   SRC_OVER
-  */
-}
