+++ src/main/java/net/rptools/maptool/model/Path.java
@@ -21,6 +21,7 @@ import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.client.ui.zone.ZoneRenderer.SelectionSet;
 import net.rptools.maptool.client.walker.NaiveWalker;
+import net.rptools.maptool.model.grid.Grid;
 import net.rptools.maptool.server.proto.PathDto;
 import net.rptools.maptool.server.proto.drawing.IntPointDto;
