+++ src/main/java/net/rptools/maptool/model/grid/SquareGrid.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/> and specifically the Affero license
  * text at <http://www.gnu.org/licenses/agpl.html>.
  */
-package net.rptools.maptool.model;
+package net.rptools.maptool.model.grid;
 
 import java.awt.Color;
 import java.awt.Dimension;
@@ -43,6 +43,10 @@ import net.rptools.maptool.client.ui.zone.ZoneRenderer;
 import net.rptools.maptool.client.walker.WalkerMetric;
 import net.rptools.maptool.client.walker.ZoneWalker;
 import net.rptools.maptool.client.walker.astar.AStarSquareEuclideanWalker;
+import net.rptools.maptool.model.CellPoint;
+import net.rptools.maptool.model.MovementKey;
+import net.rptools.maptool.model.TokenFootprint;
+import net.rptools.maptool.model.ZonePoint;
 import net.rptools.maptool.server.proto.GridDto;
 import net.rptools.maptool.server.proto.SquareGridDto;
